requirements.txt
vgamepad
pillow
pywin32
numpy
opencv-python
jax[cpu]
